[test][rfct] move oai tests into a dedicated test_resolver_oai.py
