+++ tests/test_resolver_oai.py
@@ -1,0 +1,95 @@
+from unittest import mock
+from pytest import fixture
+from shutil import copy
+from os.path import join, dirname
+
+from tests.base import main
+
+from ocrd.resolver import Resolver, extract_mets
+
+@fixture(name="response_dir")
+def fixture_response_dir(tmpdir):
+    src = './tests/data/response/oai_get_record_2200909.xml'
+    target_file = str(tmpdir.mkdir('responses').join('oai_get_record_2200909.xml'))
+    copy(src, target_file)
+    src2 = './tests/data/response/mets_kant_aufklaerung_1784.xml'
+    target_file2 = str(tmpdir.join('responses').join('mets_kant_aufklaerung_1784.xml'))
+    copy(src2, target_file2)
+    return dirname(target_file)
+
+
+@fixture(name="oai_response_content")
+def fixture_oai_2200909_content(response_dir):
+    data_path = join(response_dir, 'oai_get_record_2200909.xml')
+    with open(data_path, 'rb') as f:
+        return f.read()
+
+@fixture(name="plain_xml_response_content")
+def fixture_xml_kant_content(response_dir):
+    data_path = join(response_dir, 'mets_kant_aufklaerung_1784.xml')
+    with open(data_path, 'rb') as f:
+        return f.read()
+
+
+def test_extract_mets_from_oai_content(oai_response_content):
+    """Ensure that OAI-prelude gets dropped"""
+
+    result = extract_mets(oai_response_content)
+    expected_start = b'<?xml version="1.0" encoding="UTF-8"?>\r\n<mets:mets'
+    assert result.startswith(expected_start)
+    assert b'OAI-PHM' not in result
+
+
+
+def test_handle_response_mets(plain_xml_response_content):
+    """Ensure plain XML/Text Response is not broken"""
+
+    result = extract_mets(plain_xml_response_content)
+    expected_start = b'<?xml version="1.0"'
+    assert result.startswith(expected_start)
+
+
+@mock.patch("requests.get")
+def test_handle_common_oai_response(mock_get, response_dir, oai_response_content):
+    """Base use case with valid OAI Response data"""
+
+    # arrange
+    url = 'http://digital.bibliothek.uni-halle.de/hd/oai/?verb=GetRecord&metadataPrefix=mets&mode=xml&identifier=9049'
+    mock_get.return_value.status_code = 200
+    mock_get.return_value.content = oai_response_content
+    headers = {'Content-Type': 'text/xml'}
+    mock_get.return_value.headers = headers
+    resolver = Resolver()
+
+    # act
+    result = resolver.download_to_directory(response_dir, url)
+
+    # assert
+    mock_get.assert_called_once_with(url)
+    assert result == 'oai'
+
+
+@mock.patch("requests.get")
+@mock.patch("ocrd.resolver.log.warning")
+def test_handle_response_for_invalid_content(mock_log_warning, mock_get, response_dir):
+    """If invalid content is returned, store warning log entry"""
+
+    # arrange
+    url = 'http://digital.bibliothek.uni-halle.de/hd/oai/?verb=GetRecord&metadataPrefix=mets&mode=xml&identifier=foo'
+    mock_get.return_value.status_code = 200
+    mock_get.return_value.content = b'foo bar'
+    headers = {'Content-Type': 'text/plain'}
+    mock_get.return_value.headers = headers
+    resolver = Resolver()
+
+    # act
+    resolver.download_to_directory(response_dir, url)
+
+    # assert behavior
+    mock_get.assert_called_once_with(url)
+    assert mock_log_warning.call_count == 1
+
+
+
+if __name__ == '__main__':
+    main(__file__)
